Fixes #78: ignored properties defaultTarget and minification
The defaultTarget property was forced to 'source' by its default value in the
yargs configuration. The implementation has now been changed. The default
value, if not provided either in the defaultTarget setting in the compile.json
file, or on the command line, is to use the first target listed in compile.json.

The ignored minification was a documentation error. The correct key is `minify`.

This commit also adds emacs backup files to the list of ignored files in
.gitignore.